Separate tests for preload
